backend/requirements.txt
asgiref==3.3.1
Django==3.1.3
pytz==2020.4
sqlparse==0.4.1
djangorestframework
django-cors-headers
gunicorn
pygments
psycopg2
markdown
django-filter
drf-yasg
Pillow
pytest-django
django-rest-knox